+++ Views/DepartementView/filtre_departement.php
@@ -79,7 +79,7 @@
                         <td><?php echo $row['DepartementARAB']; ?></td>
                         <td>
                             <a href="edit_departement.php?CodeDep=<?php echo $row['CodeDep']; ?>" class="btn btn-primary">Edit</a>
-                            <a href="delete_departement.php?CodeDep=<?php echo $row['CodeDep']; ?>" class="btn btn-danger">Delete</a>
+                            <a href="delete_departement.php?CodeDep=<?php echo $row['CodeDep']; ?>" class="btn btn-danger" onclick="return confirm('Are you sure you want to delete this departement?');">Delete</a>
                         </td>
                     </tr>
                 <?php } ?>
