+++ Views/EtudiantView/filtre_etudiant.php
@@ -295,7 +295,7 @@ else{
                     <td><?php echo $row["Redaut"] ?? "vide"; ?></td>
 
                     <td><a class="btn btn-warning" href="edit_etudiant.php?NCIN=<?php echo $row["NCIN"]; ?>">Edit</a></td>
-                    <td><a class="btn btn-danger" href="delete_etudiant.php?NCIN=<?php echo $row["NCIN"]; ?>">Delete</a></td>
+                    <td><a class="btn btn-danger" href="delete_etudiant.php?NCIN=<?php echo $row["NCIN"]; ?>" onclick="return confirm('Are you sure you want to delete this Etudiant?');">Delete</a></td>
                 </tr>
             <?php endforeach; ?>
         </tbody>
