+++ Views/JourView/list_jours.php
@@ -62,7 +62,7 @@ $jourList = $jour->getJours();
                     <td><?php echo $jour['Code Prof']; ?></td>
                     <td>
                         <a href="edit_jour.php?N°=<?php echo $jour['N°']; ?>" class="btn btn-primary">Edit</a>
-                        <a href="delete_jour.php?N°=<?php echo $jour['N°']; ?>" class="btn btn-danger">Delete</a>
+                        <a href="delete_jour.php?N°=<?php echo $jour['N°']; ?>" class="btn btn-danger" onclick="return confirm(`Are you sure you want to delete this jour?`);">Delete</a>
                     </td>
                 </tr>
             <?php } ?>
