+++ Views/ProfView/filtre_prof.php
@@ -168,7 +168,7 @@ if ($_SERVER["REQUEST_METHOD"] === "POST") {
                             <td><?php echo $row['ControlAcces']; ?></td>
                             <td>
                                 <a href="edit_prof.php?Matricule=<?php echo $row['Matricule']; ?>" class="btn btn-primary">Edit</a>
-                                <a href="delete_prof.php?Matricule=<?php echo $row['Matricule']; ?>" class="btn btn-danger">Delete</a>
+                                <a href="delete_prof.php?Matricule=<?php echo $row['Matricule']; ?>" class="btn btn-danger" onclick="return confirm(`Are you sure you want to delete this Professeur?`);">Delete</a>
                             </td>
                         </tr>
                 <?php endwhile;
