+++ Views/RepartitionView/filtre_repartitions.php
@@ -118,7 +118,7 @@ if ($_SERVER["REQUEST_METHOD"] == "POST") {
                     <td><?php echo $rep['TypeGest']; ?></td>
                     <td>
                         <a href="edit_repartition.php?Numdist=<?php echo $rep['Numdist']; ?>" class="btn btn-primary">Edit</a>
-                        <a href="delete_repartition.php?Numdist=<?php echo $rep['Numdist']; ?>" class="btn btn-danger">Delete</a>
+                        <a href="delete_repartition.php?Numdist=<?php echo $rep['Numdist']; ?>" class="btn btn-danger" onclick="return confirm(`Are you sure you want to delete this repartition?`);">Delete</a>
                     </td>
                 </tr>
             <?php } ?>
